requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==5.5
CTkMessagebox==2.7
customtkinter==5.2.2
darkdetect==0.8.0
decorator==4.4.2
deprecation==2.1.0
eyed3==0.9.7
ffmpeg-python==0.2.0
filetype==1.2.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.2
packaging==24.1
pillow==11.0.0
proglog==0.1.10
pytubefix==8.1.1
requests==2.32.3
toml==0.10.2
tqdm==4.66.5
urllib3==2.2.3
